Finish the physics engine structure
The TL;DR: is, each EntityController is a PhysicsController. I need to
determine how the world fits into this. Maybe make the world itself a
PhysicsController? This is also a semi-quadtree. To be determined: what
to do with the default list of contained entities.

The current design hasn't entirely fit the world in yet, or for that
matter implemented any actual physics or collision detection, but the
way it works accounts for everything when implemented.

Essentially, each scene (not implemented either) has a set of
controllers, which are stored in a vector which creates an implicit
priority. The iteration order then indicates what's priortised. If the
game scene has a world, a set of entities, and the UI/HUD.

The order can then be:
1. UI
2. Entities (may be several controllers)
3. World

Because the UI is always on top, so it makes sense for that to be
prioritized. Secondly, there are the entities, which are on top of the
surface (except for walls, but entities aren't allowed into these tiles
anyway, so it's not a problem). Finally, if there's no UI elements
covering the area, and there are no entities on the ground, the world
has to be matched. Some tile inside the last controller has to be
clicked, which may or may not be an operation.

The details behind actually using this needs to be determined, but it
should be a working structure.

Might need to refactor it later, not sure if it's optimal but I've spent
enough time stuck on the physics engine design at this point.