Finish raytracing for the day.
Not fully functional, needs more work on the actual casting.

I've acquired the tracing vector, the rest is down to actually casting
it. Should be fairly trivial with the current system.

Might want to migrate to world coords instead of OpenGL coords;
currently a problem for text. Not sure where one bug begins and another
one ends, though. More work is needed to get there